Moved MultiChatUtil to common package
