+++ multichat/src/main/java/xyz/olivermartin/multichat/bungee/Bulletins.java
@@ -10,6 +10,7 @@ import net.md_5.bungee.api.chat.TextComponent;
 import net.md_5.bungee.api.connection.ProxiedPlayer;
 import net.md_5.bungee.api.scheduler.ScheduledTask;
 import xyz.olivermartin.multichat.bungee.events.PostBroadcastEvent;
+import xyz.olivermartin.multichat.common.MultiChatUtil;
 
 /**
  * Bulletins Management
