+++ multichat/src/main/java/xyz/olivermartin/multichat/bungee/ConsoleManager.java
@@ -7,6 +7,7 @@ import net.md_5.bungee.api.ChatColor;
 import net.md_5.bungee.api.ProxyServer;
 import net.md_5.bungee.api.chat.BaseComponent;
 import net.md_5.bungee.api.chat.TextComponent;
+import xyz.olivermartin.multichat.common.MultiChatUtil;
 
 public class ConsoleManager {
