+++ multichat/src/main/java/xyz/olivermartin/multichat/bungee/MessageManager.java
@@ -7,6 +7,7 @@ import net.md_5.bungee.api.ChatColor;
 import net.md_5.bungee.api.CommandSender;
 import net.md_5.bungee.api.chat.TextComponent;
 import net.md_5.bungee.config.Configuration;
+import xyz.olivermartin.multichat.common.MultiChatUtil;
 
 /**
  * Message Manager
