+++ multichat/src/main/java/xyz/olivermartin/multichat/bungee/commands/AnnouncementCommand.java
@@ -10,7 +10,7 @@ import net.md_5.bungee.api.plugin.Command;
 import xyz.olivermartin.multichat.bungee.Announcements;
 import xyz.olivermartin.multichat.bungee.ConfigManager;
 import xyz.olivermartin.multichat.bungee.MessageManager;
-import xyz.olivermartin.multichat.bungee.MultiChatUtil;
+import xyz.olivermartin.multichat.common.MultiChatUtil;
 
 /**
  * Announcement Command
