+++ multichat/src/main/java/xyz/olivermartin/multichat/bungee/commands/HelpMeCommand.java
@@ -10,7 +10,7 @@ import xyz.olivermartin.multichat.bungee.ChatControl;
 import xyz.olivermartin.multichat.bungee.ConfigManager;
 import xyz.olivermartin.multichat.bungee.ConsoleManager;
 import xyz.olivermartin.multichat.bungee.MessageManager;
-import xyz.olivermartin.multichat.bungee.MultiChatUtil;
+import xyz.olivermartin.multichat.common.MultiChatUtil;
 
 /**
  * 'Help Me' Command
