+++ multichat/src/main/java/xyz/olivermartin/multichat/common/MultiChatUtil.java
@@ -1,4 +1,4 @@
-package xyz.olivermartin.multichat.bungee;
+package xyz.olivermartin.multichat.common;
 
 import java.util.ArrayList;
 import java.util.Collection;
@@ -7,7 +7,7 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 public class MultiChatUtil {
-	
+
 	/**
 	 * Reformat the RGB codes into Spigot Native version
 	 * 
@@ -18,67 +18,67 @@ public class MultiChatUtil {
 		// Translate RGB codes
 		return message.replaceAll("(?i)\\&(x|#)([0-9A-F])([0-9A-F])([0-9A-F])([0-9A-F])([0-9A-F])([0-9A-F])", "&x&$2&$3&$4&$5&$6&$7");
 	}
-	
+
 	public static String approximateHexCodes(String message) {
 
-        message = message.replaceAll("(?i)(\\&|§)x(\\&|§)([0-9A-F])(\\&|§)([0-9A-F])(\\&|§)([0-9A-F])(\\&|§)([0-9A-F])(\\&|§)([0-9A-F])(\\&|§)([0-9A-F])", "&#$3$5$7$9$11$13");
+		message = message.replaceAll("(?i)(\\&|§)x(\\&|§)([0-9A-F])(\\&|§)([0-9A-F])(\\&|§)([0-9A-F])(\\&|§)([0-9A-F])(\\&|§)([0-9A-F])(\\&|§)([0-9A-F])", "&#$3$5$7$9$11$13");
 
-        List<String> allMatches = new ArrayList<String>();
-        Matcher m = Pattern.compile("(?i)\\&(x|#)([0-9A-F])([0-9A-F])([0-9A-F])([0-9A-F])([0-9A-F])([0-9A-F])")
-            .matcher(message);
-        while (m.find()) {
-            allMatches.add(m.group());
-        }
+		List<String> allMatches = new ArrayList<String>();
+		Matcher m = Pattern.compile("(?i)\\&(x|#)([0-9A-F])([0-9A-F])([0-9A-F])([0-9A-F])([0-9A-F])([0-9A-F])")
+				.matcher(message);
+		while (m.find()) {
+			allMatches.add(m.group());
+		}
 
-        for (String match : allMatches) {
-            String hexonly = match.split("#")[1];
-            String minecraftCode = hexToMinecraft(hexonly);
-            message = message.replace(match,"§"+minecraftCode);
-        }
+		for (String match : allMatches) {
+			String hexonly = match.split("#")[1];
+			String minecraftCode = hexToMinecraft(hexonly);
+			message = message.replace(match,"§"+minecraftCode);
+		}
 
-        return message;
+		return message;
 
-    }
+	}
 
-    public static String hexToMinecraft(String hex) {
+	public static String hexToMinecraft(String hex) {
 
-        String rcode = hex.substring(0,2);
-        String gcode = hex.substring(2,4);
-        String bcode = hex.substring(4,6);
+		String rcode = hex.substring(0,2);
+		String gcode = hex.substring(2,4);
+		String bcode = hex.substring(4,6);
 
-        int rint = Integer.parseInt(rcode,16);
-        int gint = Integer.parseInt(gcode,16);
-        int bint = Integer.parseInt(bcode,16);
+		int rint = Integer.parseInt(rcode,16);
+		int gint = Integer.parseInt(gcode,16);
+		int bint = Integer.parseInt(bcode,16);
 
-        String[] cga = {"000000","0000aa","00aa00","00aaaa","aa0000","aa00aa","ffaa00","aaaaaa","555555","5555ff","55ff55","55ffff","ff5555","ff55ff","ffff55","ffffff"};
+		String[] cga = {"000000","0000aa","00aa00","00aaaa","aa0000","aa00aa","ffaa00","aaaaaa","555555","5555ff","55ff55","55ffff","ff5555","ff55ff","ffff55","ffffff"};
 
-        int diff = 999999999;
-        int best = -1;
+		int diff = 999999999;
+		int best = -1;
 
-        for (int i = 0; i < 16; i++) {
+		for (int i = 0; i < 16; i++) {
 
-            String current = cga[i];
+			String current = cga[i];
 
-            String rcode2 = current.substring(0,2);
-            String gcode2 = current.substring(2,4);
-            String bcode2 = current.substring(4,6);
+			String rcode2 = current.substring(0,2);
+			String gcode2 = current.substring(2,4);
+			String bcode2 = current.substring(4,6);
 
-            int rint2 = Integer.parseInt(rcode2,16);
-            int gint2 = Integer.parseInt(gcode2,16);
-            int bint2 = Integer.parseInt(bcode2,16);
+			int rint2 = Integer.parseInt(rcode2,16);
+			int gint2 = Integer.parseInt(gcode2,16);
+			int bint2 = Integer.parseInt(bcode2,16);
 
-            int val = Math.abs(rint-rint2) + Math.abs(gint-gint2) + Math.abs(bint-bint2);
+			int val = Math.abs(rint-rint2) + Math.abs(gint-gint2) + Math.abs(bint-bint2);
 
-            if (val < diff) {
-                best = i;
-                diff = val;
-            }
+			if (val < diff) {
+				best = i;
+				diff = val;
+			}
 
-        }
+		}
 
-        return Integer.toHexString(best);
+		return Integer.toHexString(best);
 
-    }
+	}
 
 	/**
 	 * Concatenate the arguments together to get the message as a string
